Refactor CaptureScreen to isolate P/Invokes
